+++ src/main/kotlin/io/github/gunpowder/signtypes/AdminBuySign.kt
@@ -24,14 +24,15 @@
 
 package io.github.gunpowder.signtypes
 
-import io.github.gunpowder.GunpowderSignshopModule
 import io.github.gunpowder.api.builders.SignType
 import io.github.gunpowder.api.builders.Text
 import io.github.gunpowder.api.components.with
 import io.github.gunpowder.entities.ConfirmPopup
+import io.github.gunpowder.entities.SignAdminBuyData
+import io.github.gunpowder.entities.SignDataComponent
 import io.github.gunpowder.entities.SignPlayerComponent
 import io.github.gunpowder.modelhandlers.BalanceHandler
-import net.minecraft.block.entity.SignBlockEntity
+import net.minecraft.entity.ItemEntity
 import net.minecraft.item.ItemStack
 import net.minecraft.nbt.NbtCompound
 import net.minecraft.text.LiteralText
@@ -40,7 +41,6 @@ import net.minecraft.util.ItemScatterer
 import net.minecraft.util.collection.DefaultedList
 
 object AdminBuySign {
-    val dataCache = mutableMapOf<SignBlockEntity, SignBuyData>()
     val handler by lazy {
         BalanceHandler
     }
@@ -52,7 +52,8 @@ object AdminBuySign {
             requires { signBlockEntity, serverPlayerEntity -> serverPlayerEntity.hasPermissionLevel(4) }
 
             onClicked { signBlockEntity, serverPlayerEntity ->
-                val data = dataCache[signBlockEntity] ?: return@onClicked
+                val comp = signBlockEntity.with<SignDataComponent<SignAdminBuyData>>()
+                val data = comp.data ?: return@onClicked
 
                 ConfirmPopup(Text.builder {
                     text("Buying ${data.targetStack} for $${data.price}. ")
@@ -70,8 +71,9 @@ object AdminBuySign {
                             it
                         }
 
-                        if (!serverPlayerEntity.inventory.insertStack(data.targetStack.copy())) {
-                            ItemScatterer.spawn(serverPlayerEntity.world, serverPlayerEntity.blockPos, DefaultedList.copyOf(data.targetStack.copy()))
+                        val stack = data.targetStack.copy()
+                        if (!serverPlayerEntity.inventory.insertStack(stack)) {
+                            serverPlayerEntity.world.spawnEntity(ItemEntity(serverPlayerEntity.world, serverPlayerEntity.x, serverPlayerEntity.y, serverPlayerEntity.z, stack))
                         }
                         serverPlayerEntity.sendMessage(LiteralText("Purchased ${data.targetStack} for $${data.price}"), false)
                     }
@@ -93,9 +95,9 @@ object AdminBuySign {
                             signBlockEntity.world?.removeBlock(signBlockEntity.pos, false)
                         } else {
                             serverPlayerEntity.sendMessage(LiteralText("Put up for sale: $stack for $${price.first()!!}"), false)
-                            val data = SignBuyData(stack.copy(), price.first()!!)
+                            val data = SignAdminBuyData(stack.copy(), price.first()!!)
                             comp.selected = null
-                            dataCache[signBlockEntity] = data
+                            signBlockEntity.with<SignDataComponent<SignAdminBuyData>>().data = data
                         }
                     }
                 } else {
@@ -105,13 +107,13 @@ object AdminBuySign {
             }
 
             onDestroyed { signBlockEntity, serverPlayerEntity ->
-                if (dataCache.containsKey(signBlockEntity)) {
-                    dataCache.remove(signBlockEntity)
-                }
+                signBlockEntity.with<SignDataComponent<SignAdminBuyData>>().data = null
+
             }
 
             serialize { signBlockEntity, compoundTag ->
-                val data = dataCache[signBlockEntity]  ?: return@serialize
+                val comp = signBlockEntity.with<SignDataComponent<SignAdminBuyData>>()
+                val data = comp.data ?: return@serialize
                 val link = NbtCompound()
                 val item = NbtCompound()
                 data.targetStack.writeNbt(item)
@@ -122,17 +124,12 @@ object AdminBuySign {
 
             deserialize { signBlockEntity, compoundTag ->
                 val link = compoundTag.getCompound("link")
-                val data = SignBuyData(
+                val comp = signBlockEntity.with<SignDataComponent<SignAdminBuyData>>()
+                comp.data = SignAdminBuyData(
                     ItemStack.fromNbt(link.getCompound("item")),
                     link.getDouble("price")
                 )
-                dataCache[signBlockEntity] = data
             }
         }
     }
-
-    data class SignBuyData(
-        val targetStack: ItemStack,
-        val price: Double
-    )
 }